+++ dart/collision/native/fwd.hpp
@@ -80,6 +80,8 @@ struct BatchTimings;
 struct BatchStats;
 struct BatchOutput;
 struct BroadPhaseSnapshot;
+struct BroadPhaseDebugCell;
+struct BroadPhaseDebugEndpoint;
 struct BroadPhaseDebugNode;
 struct BroadPhaseDebugSnapshot;